requirements.txt
keyboard>=0.13.5
pynput>=1.7.6
pystray>=0.19.4
Pillow>=10.0.0
pywin32>=306
openai>=1.30.0
anthropic>=0.25.0
